+++ packages/gamut-styles/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [17.12.1](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-styles@17.12.0...@codecademy/gamut-styles@17.12.1) (2026-03-17)
+
+### Bug Fixes
+
+- update reboot file to conditionally use logical properties or physical properties ([404937b](https://github.com/Codecademy/gamut/commit/404937be816a16ffbf7cb4c5b3a2a1999969f31d))
+
+## [17.12.0](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-styles@17.11.5...@codecademy/gamut-styles@17.12.0) (2026-03-12)
+
+### Features
+
+- Logical properties ([4c340c3](https://github.com/Codecademy/gamut/commit/4c340c3949bad30fce4f9ffabb2018763df5c4ba))
+
 ### [17.11.5](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-styles@17.11.4...@codecademy/gamut-styles@17.11.5) (2026-03-04)
 
 ### Bug Fixes