+++ packages/gamut/CHANGELOG.md
@@ -3,6 +3,24 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [68.1.3](https://github.com/Codecademy/gamut/compare/@codecademy/gamut@68.1.2...@codecademy/gamut@68.1.3) (2026-03-18)
+
+**Note:** Version bump only for package @codecademy/gamut
+
+### [68.1.2](https://github.com/Codecademy/gamut/compare/@codecademy/gamut@68.1.1...@codecademy/gamut@68.1.2) (2026-03-17)
+
+**Note:** Version bump only for package @codecademy/gamut
+
+### [68.1.1](https://github.com/Codecademy/gamut/compare/@codecademy/gamut@68.1.0...@codecademy/gamut@68.1.1) (2026-03-17)
+
+**Note:** Version bump only for package @codecademy/gamut
+
+## [68.1.0](https://github.com/Codecademy/gamut/compare/@codecademy/gamut@68.0.5...@codecademy/gamut@68.1.0) (2026-03-12)
+
+### Features
+
+- Logical properties ([4c340c3](https://github.com/Codecademy/gamut/commit/4c340c3949bad30fce4f9ffabb2018763df5c4ba))
+
 ### [68.0.5](https://github.com/Codecademy/gamut/compare/@codecademy/gamut@68.0.4...@codecademy/gamut@68.0.5) (2026-03-04)
 
 ### Bug Fixes